Added 8th Button, Removed SDA_IN
SDA_IN not needed for I2C to function (as display is write only) so it has been removed to save a pin. In the future if needed, the existing SDA_OUT pin can be reconfig'ed into an inout pin anyway. Added the chassis button S2 as the 8th button. Will add to docs in later update